splitted reducer
