+++ project/src/components/private-route/private-route.jsx
@@ -25,8 +25,8 @@ PrivateRoute.propTypes = {
   authorizationStatus: PropTypes.string.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
+const mapStateToProps = ({AUTH}) => ({
+  authorizationStatus: AUTH.authorizationStatus,
 });
 
 export {PrivateRoute};
